creat kobject package, make loader and transformer refering to it
