+++ cli/app/app.go
@@ -17,103 +17,36 @@ limitations under the License.
 package app
 
 import (
-	"fmt"
-	"math/rand"
-	"os"
-	"path/filepath"
-	"strconv"
-	"strings"
-
 	"github.com/Sirupsen/logrus"
 	"github.com/urfave/cli"
-
-	"encoding/json"
-	"io/ioutil"
 
 	// install kubernetes api
 	_ "k8s.io/kubernetes/pkg/api/install"
 	_ "k8s.io/kubernetes/pkg/apis/extensions/install"
 
 	"k8s.io/kubernetes/pkg/api"
-	"k8s.io/kubernetes/pkg/api/unversioned"
 	"k8s.io/kubernetes/pkg/apis/extensions"
 	client "k8s.io/kubernetes/pkg/client/unversioned"
 	cmdutil "k8s.io/kubernetes/pkg/kubectl/cmd/util"
 
-	"k8s.io/kubernetes/pkg/runtime"
-	"k8s.io/kubernetes/pkg/util/intstr"
-
-	deployapi "github.com/openshift/origin/pkg/deploy/api"
 	// install kubernetes api
 	_ "github.com/openshift/origin/pkg/deploy/api/install"
 
-	"github.com/ghodss/yaml"
 	"github.com/skippbox/kompose/pkg/kobject"
+	"github.com/skippbox/kompose/pkg/loader"
 	"github.com/skippbox/kompose/pkg/transformer"
 	"github.com/docker/libcompose/lookup"
 	"github.com/docker/libcompose/config"
 	"github.com/docker/libcompose/project"
+	"fmt"
+	"strings"
 )
 
 const (
-	letterBytes        = "abcdefghijklmnopqrstuvwxyz0123456789"
 	DefaultComposeFile = "docker-compose.yml"
 )
 
 var inputFormat = "compose"
-
-var composeOptions = map[string]string{
-	"Build":         "build",
-	"CapAdd":        "cap_add",
-	"CapDrop":       "cap_drop",
-	"CPUSet":        "cpuset",
-	"CPUShares":     "cpu_shares",
-	"CPUQuota":      "cpu_quota",
-	"CgroupParent":  "cgroup_parent",
-	"Devices":       "devices",
-	"DependsOn":     "depends_on",
-	"DNS":           "dns",
-	"DNSSearch":     "dns_search",
-	"DomainName":    "domainname",
-	"Entrypoint":    "entrypoint",
-	"EnvFile":       "env_file",
-	"Expose":        "expose",
-	"Extends":       "extends",
-	"ExternalLinks": "external_links",
-	"ExtraHosts":    "extra_hosts",
-	"Hostname":      "hostname",
-	"Ipc":           "ipc",
-	"Logging":       "logging",
-	"MacAddress":    "mac_address",
-	"MemLimit":      "mem_limit",
-	"MemSwapLimit":  "memswap_limit",
-	"NetworkMode":   "network_mode",
-	"Networks":      "networks",
-	"Pid":           "pid",
-	"SecurityOpt":   "security_opt",
-	"ShmSize":       "shm_size",
-	"StopSignal":    "stop_signal",
-	"VolumeDriver":  "volume_driver",
-	"VolumesFrom":   "volumes_from",
-	"Uts":           "uts",
-	"ReadOnly":      "read_only",
-	"StdinOpen":     "stdin_open",
-	"Tty":           "tty",
-	"User":          "user",
-	"Ulimits":       "ulimits",
-	"Dockerfile":    "dockerfile",
-	"Net":           "net",
-	"Args":          "args",
-}
-
-// RandStringBytes generates randomly n-character string
-func RandStringBytes(n int) string {
-	b := make([]byte, n)
-	for i := range b {
-		b[i] = letterBytes[rand.Intn(len(letterBytes))]
-	}
-	return string(b)
-}
 
 // BeforeApp is an action that is executed before any cli command.
 func BeforeApp(c *cli.Context) error {
@@ -256,351 +189,6 @@ func Scale(c *cli.Context) {
 	//		fmt.Printf("Scaling %s to: %d\n", name, s.Spec.Replicas)
 	//	}
 	//}
-}
-
-// Create the file to write to if --out is specified
-func createOutFile(out string) *os.File {
-	var f *os.File
-	var err error
-	if len(out) != 0 {
-		f, err = os.Create(out)
-		if err != nil {
-			logrus.Fatalf("error opening file: %v", err)
-		}
-	}
-	return f
-}
-
-// Init RC object
-func initRC(name string, service kobject.ServiceConfig, replicas int) *api.ReplicationController {
-	rc := &api.ReplicationController{
-		TypeMeta: unversioned.TypeMeta{
-			Kind:       "ReplicationController",
-			APIVersion: "v1",
-		},
-		ObjectMeta: api.ObjectMeta{
-			Name: name,
-			//Labels: map[string]string{"service": name},
-		},
-		Spec: api.ReplicationControllerSpec{
-			Selector: map[string]string{"service": name},
-			Replicas: int32(replicas),
-			Template: &api.PodTemplateSpec{
-				ObjectMeta: api.ObjectMeta{
-				//Labels: map[string]string{"service": name},
-				},
-				Spec: api.PodSpec{
-					Containers: []api.Container{
-						{
-							Name:  name,
-							Image: service.Image,
-						},
-					},
-				},
-			},
-		},
-	}
-	return rc
-}
-
-// Init SC object
-func initSC(name string, service kobject.ServiceConfig) *api.Service {
-	sc := &api.Service{
-		TypeMeta: unversioned.TypeMeta{
-			Kind:       "Service",
-			APIVersion: "v1",
-		},
-		ObjectMeta: api.ObjectMeta{
-			Name: name,
-			//Labels: map[string]string{"service": name},
-		},
-		Spec: api.ServiceSpec{
-			Selector: map[string]string{"service": name},
-		},
-	}
-	return sc
-}
-
-// Init DC object
-func initDC(name string, service kobject.ServiceConfig, replicas int) *extensions.Deployment {
-	dc := &extensions.Deployment{
-		TypeMeta: unversioned.TypeMeta{
-			Kind:       "Deployment",
-			APIVersion: "extensions/v1beta1",
-		},
-		ObjectMeta: api.ObjectMeta{
-			Name:   name,
-			Labels: map[string]string{"service": name},
-		},
-		Spec: extensions.DeploymentSpec{
-			Replicas: int32(replicas),
-			Selector: &unversioned.LabelSelector{
-				MatchLabels: map[string]string{"service": name},
-			},
-			//UniqueLabelKey: p.Name,
-			Template: api.PodTemplateSpec{
-				ObjectMeta: api.ObjectMeta{
-					Labels: map[string]string{"service": name},
-				},
-				Spec: api.PodSpec{
-					Containers: []api.Container{
-						{
-							Name:  name,
-							Image: service.Image,
-						},
-					},
-				},
-			},
-		},
-	}
-	return dc
-}
-
-// Init DS object
-func initDS(name string, service kobject.ServiceConfig) *extensions.DaemonSet {
-	ds := &extensions.DaemonSet{
-		TypeMeta: unversioned.TypeMeta{
-			Kind:       "DaemonSet",
-			APIVersion: "extensions/v1beta1",
-		},
-		ObjectMeta: api.ObjectMeta{
-			Name: name,
-		},
-		Spec: extensions.DaemonSetSpec{
-			Template: api.PodTemplateSpec{
-				ObjectMeta: api.ObjectMeta{
-					Name: name,
-				},
-				Spec: api.PodSpec{
-					Containers: []api.Container{
-						{
-							Name:  name,
-							Image: service.Image,
-						},
-					},
-				},
-			},
-		},
-	}
-	return ds
-}
-
-// Init RS object
-func initRS(name string, service kobject.ServiceConfig, replicas int) *extensions.ReplicaSet {
-	rs := &extensions.ReplicaSet{
-		TypeMeta: unversioned.TypeMeta{
-			Kind:       "ReplicaSet",
-			APIVersion: "extensions/v1beta1",
-		},
-		ObjectMeta: api.ObjectMeta{
-			Name: name,
-		},
-		Spec: extensions.ReplicaSetSpec{
-			Replicas: int32(replicas),
-			Selector: &unversioned.LabelSelector{
-				MatchLabels: map[string]string{"service": name},
-			},
-			Template: api.PodTemplateSpec{
-				ObjectMeta: api.ObjectMeta{},
-				Spec: api.PodSpec{
-					Containers: []api.Container{
-						{
-							Name:  name,
-							Image: service.Image,
-						},
-					},
-				},
-			},
-		},
-	}
-	return rs
-}
-
-// initDeploymentConfig initialize OpenShifts DeploymentConfig object
-func initDeploymentConfig(name string, service kobject.ServiceConfig, replicas int) *deployapi.DeploymentConfig {
-	dc := &deployapi.DeploymentConfig{
-		TypeMeta: unversioned.TypeMeta{
-			Kind:       "DeploymentConfig",
-			APIVersion: "v1",
-		},
-		ObjectMeta: api.ObjectMeta{
-			Name:   name,
-			Labels: map[string]string{"service": name},
-		},
-		Spec: deployapi.DeploymentConfigSpec{
-			Replicas: int32(replicas),
-			Selector: map[string]string{"service": name},
-			//UniqueLabelKey: p.Name,
-			Template: &api.PodTemplateSpec{
-				ObjectMeta: api.ObjectMeta{
-					Labels: map[string]string{"service": name},
-				},
-				Spec: api.PodSpec{
-					Containers: []api.Container{
-						{
-							Name:  name,
-							Image: service.Image,
-						},
-					},
-				},
-			},
-		},
-	}
-	return dc
-}
-
-// Configure the environment variables.
-func configEnvs(name string, service kobject.ServiceConfig) []api.EnvVar {
-	envs := []api.EnvVar{}
-	for _, v := range service.Environment {
-		envs = append(envs, api.EnvVar{
-			Name:  v.Name,
-			Value: v.Value,
-		})
-	}
-
-	return envs
-}
-
-// Configure the container volumes.
-func configVolumes(service kobject.ServiceConfig) ([]api.VolumeMount, []api.Volume) {
-	volumesMount := []api.VolumeMount{}
-	volumes := []api.Volume{}
-	volumeSource := api.VolumeSource{}
-	for _, volume := range service.Volumes {
-		name, host, container, mode, err := parseVolume(volume)
-		if err != nil {
-			logrus.Warningf("Failed to configure container volume: %v", err)
-			continue
-		}
-
-		// if volume name isn't specified, set it to a random string of 20 chars
-		if len(name) == 0 {
-			name = RandStringBytes(20)
-		}
-		// check if ro/rw mode is defined, default rw
-		readonly := len(mode) > 0 && mode == "ro"
-
-		volumesMount = append(volumesMount, api.VolumeMount{Name: name, ReadOnly: readonly, MountPath: container})
-
-		if len(host) > 0 {
-			volumeSource = api.VolumeSource{HostPath: &api.HostPathVolumeSource{Path: host}}
-		} else {
-			volumeSource = api.VolumeSource{EmptyDir: &api.EmptyDirVolumeSource{}}
-		}
-
-		volumes = append(volumes, api.Volume{Name: name, VolumeSource: volumeSource})
-	}
-	return volumesMount, volumes
-}
-
-// parseVolume parse a given volume, which might be [name:][host:]container[:access_mode]
-func parseVolume(volume string) (name, host, container, mode string, err error) {
-	separator := ":"
-	volumeStrings := strings.Split(volume, separator)
-	if len(volumeStrings) == 0 {
-		return
-	}
-	// Set name if existed
-	if !isPath(volumeStrings[0]) {
-		name = volumeStrings[0]
-		volumeStrings = volumeStrings[1:]
-	}
-	if len(volumeStrings) == 0 {
-		err = fmt.Errorf("invalid volume format: %s", volume)
-		return
-	}
-	if volumeStrings[len(volumeStrings)-1] == "rw" || volumeStrings[len(volumeStrings)-1] == "ro" {
-		mode = volumeStrings[len(volumeStrings)-1]
-		volumeStrings = volumeStrings[:len(volumeStrings)-1]
-	}
-	container = volumeStrings[len(volumeStrings)-1]
-	volumeStrings = volumeStrings[:len(volumeStrings)-1]
-	if len(volumeStrings) == 1 {
-		host = volumeStrings[0]
-	}
-	if !isPath(container) || (len(host) > 0 && !isPath(host)) || len(volumeStrings) > 1 {
-		err = fmt.Errorf("invalid volume format: %s", volume)
-		return
-	}
-	return
-}
-
-func isPath(substring string) bool {
-	return strings.Contains(substring, "/")
-}
-
-// Configure the container ports.
-func configPorts(name string, service kobject.ServiceConfig) []api.ContainerPort {
-	ports := []api.ContainerPort{}
-	for _, port := range service.Port {
-		var p api.Protocol
-		switch port.Protocol {
-		default:
-			p = api.ProtocolTCP
-		case kobject.ProtocolTCP:
-			p = api.ProtocolTCP
-		case kobject.ProtocolUDP:
-			p = api.ProtocolUDP
-		}
-		ports = append(ports, api.ContainerPort{
-			ContainerPort: port.ContainerPort,
-			Protocol:      p,
-		})
-	}
-
-	return ports
-}
-
-// Configure the container service ports.
-func configServicePorts(name string, service kobject.ServiceConfig) []api.ServicePort {
-	servicePorts := []api.ServicePort{}
-	for _, port := range service.Port {
-		if port.HostPort == 0 {
-			port.HostPort = port.ContainerPort
-		}
-		var p api.Protocol
-		switch port.Protocol {
-		default:
-			p = api.ProtocolTCP
-		case kobject.ProtocolTCP:
-			p = api.ProtocolTCP
-		case kobject.ProtocolUDP:
-			p = api.ProtocolUDP
-		}
-		var targetPort intstr.IntOrString
-		targetPort.IntVal = port.ContainerPort
-		targetPort.StrVal = strconv.Itoa(int(port.ContainerPort))
-		servicePorts = append(servicePorts, api.ServicePort{
-			Name:       strconv.Itoa(int(port.HostPort)),
-			Protocol:   p,
-			Port:       port.HostPort,
-			TargetPort: targetPort,
-		})
-	}
-	return servicePorts
-}
-
-// Transform data to json/yaml
-func transformData(obj runtime.Object, generateYaml bool) ([]byte, error) {
-	//  Convert to versioned object
-	objectVersion := obj.GetObjectKind().GroupVersionKind()
-	version := unversioned.GroupVersion{Group: objectVersion.Group, Version: objectVersion.Version}
-	versionedObj, err := api.Scheme.ConvertToVersion(obj, version)
-	if err != nil {
-		return nil, err
-	}
-
-	// convert data to json / yaml
-	data, err := json.MarshalIndent(versionedObj, "", "  ")
-	if generateYaml == true {
-		data, err = yaml.Marshal(versionedObj)
-	}
-	if err != nil {
-		return nil, err
-	}
-	logrus.Debugf("%s\n", data)
-	return data, nil
 }
 
 // Convert komposeObject to K8S controllers
@@ -861,6 +449,17 @@ func Convert(c *cli.Context) {
 		file = dabFile
 	}
 
+	//komposeObject.Loader(file, inputFormat)
+	switch inputFormat {
+	case "bundle":
+		komposeObject = loader.LoadBundle(file)
+	case "compose":
+		komposeObject = loader.LoadCompose(file)
+	default:
+		logrus.Fatalf("Input file format is not supported")
+
+	}
+
 	opt := kobject.ConvertOptions{
 		ToStdout:               toStdout,
 		CreateD:                createD,
@@ -874,47 +473,13 @@ func Convert(c *cli.Context) {
 		OutFile:                outFile,
 	}
 
-	f := createOutFile(opt.OutFile)
-	defer f.Close()
-
 	validateFlags(opt, singleOutput, dabFile, inputFile)
 
-	komposeObject.Loader(file, inputFormat)
-
-	//komposeObject.Transformer(opt)
 	// Convert komposeObject to K8S controllers
 	mServices, mDeployments, mDaemonSets, mReplicationControllers, mDeploymentConfigs, svcnames := transformer.Transform(komposeObject, opt)
 
 	// Print output
 	transformer.PrintControllers(mServices, mDeployments, mDaemonSets, mReplicationControllers, mDeploymentConfigs, svcnames, opt, f)
-}
-
-// Either print to stdout or to file/s
-func print(name, trailing string, data []byte, toStdout, generateYaml bool, f *os.File) string {
-
-	file := ""
-	if generateYaml {
-		file = fmt.Sprintf("%s-%s.yaml", name, trailing)
-	} else {
-		file = fmt.Sprintf("%s-%s.json", name, trailing)
-	}
-	if toStdout {
-		fmt.Fprintf(os.Stdout, "%s\n", string(data))
-		return ""
-	} else if f != nil {
-		// Write all content to a single file f
-		if _, err := f.WriteString(fmt.Sprintf("%s\n", string(data))); err != nil {
-			logrus.Fatalf("Failed to write %s to file: %v", trailing, err)
-		}
-		f.Sync()
-	} else {
-		// Write content separately to each file
-		if err := ioutil.WriteFile(file, []byte(data), 0644); err != nil {
-			logrus.Fatalf("Failed to write %s: %v", trailing, err)
-		}
-		logrus.Printf("file %q created", file)
-	}
-	return file
 }
 
 // Up brings up deployment, svc.
@@ -987,28 +552,4 @@ func sortServicesFirst(objs []runtime.Object) []runtime.Object {
 	ret = append(ret, svc...)
 	ret = append(ret, others...)
 	return ret
-}
-
-// updateController updates the given object with the given pod template update function and ObjectMeta update function
-func updateController(obj runtime.Object, updateTemplate func(*api.PodTemplateSpec), updateMeta func(meta *api.ObjectMeta)) {
-	switch t := obj.(type) {
-	case *api.ReplicationController:
-		if t.Spec.Template == nil {
-			t.Spec.Template = &api.PodTemplateSpec{}
-		}
-		updateTemplate(t.Spec.Template)
-		updateMeta(&t.ObjectMeta)
-	case *extensions.Deployment:
-		updateTemplate(&t.Spec.Template)
-		updateMeta(&t.ObjectMeta)
-	case *extensions.ReplicaSet:
-		updateTemplate(&t.Spec.Template)
-		updateMeta(&t.ObjectMeta)
-	case *extensions.DaemonSet:
-		updateTemplate(&t.Spec.Template)
-		updateMeta(&t.ObjectMeta)
-	case *deployapi.DeploymentConfig:
-		updateTemplate(t.Spec.Template)
-		updateMeta(&t.ObjectMeta)
-	}
 }
